% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_namespace_gen.R
\docType{package}
\name{SampleSize4ClinicalTrials}
\alias{SampleSize4ClinicalTrials}
\alias{SampleSize4ClinicalTrials-package}
\title{Sample Size Calculation for the Comparison of Means or Proportions in Phase III Clinical Trials}
\description{
There are four categories for Phase III clinical trials according to different research goals, including (1) Testing for equality, (2) Superiority trial, (3) Non-inferiority trial, and (4) Equivalence trial. This package aims to help researchers to calculate sample size when comparing means or proportions in Phase III clinical trials with different research goals.
}
\author{
Hongchao Qi, Fang Zhu
}
